Removes most of Visual Studio warnings

+++ src/interaction/interactionhandler.cpp
@@ -313,7 +313,7 @@ int distance(lua_State* L) {
 
 	double d1 = luaL_checknumber(L, -2);
 	double d2 = luaL_checknumber(L, -1);
-	PowerScaledScalar dist(d1, d2);
+	PowerScaledScalar dist(static_cast<float>(d1), static_cast<float>(d2));
 	OsEng.interactionHandler().distanceDelta(dist);
 	return 0;
 }
@@ -515,7 +515,7 @@ void InteractionHandler::mousePositionCallback(int x, int y) {
 
 void InteractionHandler::mouseScrollWheelCallback(int pos) {
 	if (_mouseController)
-		_mouseController->scrollWheel(float(pos));
+		_mouseController->scrollWheel(pos);
 }
 
 void InteractionHandler::orbitDelta(const glm::quat& rotation)
